# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api for the e-com website",
        "description": "This Api will help in communication of multiple devices want to access our services",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3200",
            "description": "this is main url for our api"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    },
    "paths": {
...
        "/api/product": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "It create a product in the site",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "example": "Bluetooth Earbuds",
                                        "type": "string"
                                    },
                                    "desc": {
                                        "example": "Wireless Bluetooth earbuds with noise cancellation",
                                        "type": "string"
                                    },
                                    "imageUrl": {
                                        "example": "https://example.com/earbuds.jpg",
                                        "type": "string"
                                    },
                                    "category": {
                                        "example": "category3",
                                        "type": "string"
                                    },
                                    "price": {
                                        "example": 129.99,
                                        "type": "number"
                                    },
                                    "sizes": {
                                        "example": "S,L,XL",
                                        "type": "string"
                                    },
                                    "resume": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Upload the resume file"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "It requires to send the authorization header with the request",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Product is not created"
                    },
                    "401": {
                        "description": "First login"
                    }
                }
            },
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "It get all the products in the site",
                "description": "It requires to send the authorization header with the request",
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Non Product is listed"
                    },
                    "401": {
                        "description": "First login"
                    }
                }
            }
        },
...
        "/api/cart": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "security": [
                    {
                        "JWT": {}
                    }
                ],
                "summary": "Add a product to the cart",
                "description": "Allows users to add a product to their cart by providing the product ID and quantity.",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "673ba50d5a51122b10432f62"
                        },
                        "description": "The unique ID of the product to add to the cart"
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 2
                        },
                        "description": "The number of units of the product to add"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product added to cart successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Product added to cart",
                                    "productId": "673ba50d5a51122b10432f62",
                                    "quantity": 2
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request (e.g., negative quantity)"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    }
}